Move the duplicate logic to seperate keymgt service class.
